requirements.txt
h5py
numpy
scikit-learn
docker
PyYAML
matplotlib